requirements.txt
beautifulsoup4==4.11.1
markdownify==0.11.2
python-dateutil==2.8.2
sh==1.14.2
six==1.16.0
soupsieve==2.3.2.post1
